refresh events on ADD||DEL, move styles
