+++ src/assets/styles/eventCard.scss
