+++ src/assets/styles/events.scss
