+++ src/assets/styles/team.css
