+++ src/components/About.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './about.css';
+import '../assets/styles/about.css';
 
 export default function About() {
     return (
