+++ src/components/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './home.css';
+import '../assets/styles/home.css';
 
 export default function Home() {
     return (
