+++ src/components/Nav.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './nav.css';
+import '../assets/styles/nav.css';
 import { PAGE_NAME } from '../Constants';
-import SrujanaLogo from '../assets/srujana-logo.png'
+import SrujanaLogo from '../assets/images/srujana-logo.png'
 
 export default function Nav({ setPage }) {
     return (
